backend/requirements.txt

openpyxl
python-dotenv
openai>=1.40.0




fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
pandas==2.2.2
openpyxl==3.1.5
httpx==0.27.2
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.5.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.12


# UMAP/HDBSCAN need numba/llvmlite
numba>=0.59
llvmlite>=0.42

umap-learn>=0.5.5
hdbscan==0.8.37
yake>=0.4.8

# text embeddings
sentence-transformers>=2.6.0
